app/requirements.txt
alembic==1.3.1
astroid==2.3.3
Babel==2.7.0
bcrypt==3.1.7
blinker==1.4
cffi==1.13.2
Click==7.0
Flask==1.1.1
Flask-Admin==1.5.4
Flask-Babel==0.12.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-MongoAlchemy==0.7.2
Flask-SQLAlchemy==2.4.1
Flask-Table==0.5.0
Flask-WTF==0.14.2
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
MongoAlchemy==0.19
mpmath==1.1.0
Pillow==6.2.1
pycparser==2.19
pylint==2.4.4
pymongo==2.8.1
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.3
six==1.13.0
SQLAlchemy==1.3.11
sympy==1.4
typed-ast==1.4.0
Werkzeug==0.16.0
wrapt==1.11.2
WTForms==2.2.1
